packages/plugin-vercelai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neocortex-vercelai"
version = "0.1.0"
description = "Vercel AI SDK integrations for TinyHumans memory API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "tinyhumansai>=0.5",
    "ai_sdk>=0.0.1",
    "pydantic>=2.4.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
examples = ["python-dotenv>=1.0"]

[tool.setuptools.packages.find]
include = ["neocortex_vercelai*"]

[tool.ruff]
line-length = 88
target-version = "py39"
